Adds basic support for one to many associations.
This only works for List members.  I'll take care of Set and arrays shortly.

The testcases are JPA only and I still want to write more before I move on to
JDO, but this CL is already plenty large so I want to get it reviewed and
checked in before I pile on anything else.

There is a bunch of code that I once again had to borrow from the datanucleus
RDBMS adapter.