Arrays are a non-standard feature in DataNucleus and some of the use cases don't even work properly in RDBMS.  Nevertheless, some basic operations work as expected.
I also discovered that we weren't properly handling the case where someone
modifies a persistent list by setting an element at a specific location.  This
was broken for both JDO and JPA.  I suspect there is a similar problem with
remove() but I'll tackle that in a subsequent CL.

JDO support for Sets and Arrays
JPA support for Sets (JPA does allow persistet arrays).

Lots of refactoring in this CL so that I could reuse tests for List, Array, and
Set.

Also added a bunch more testcases for List.