Now that we can handle multiple updates to the same entity in a single txn we can start storing child keys on the parent.
This CL just changes how we write things and updates all existing tests
to pass - I haven't updated the read logic to take advantage of the new
fields.

There are also more tests to be written and a few easy optimizations to
avoid extraneous updates, but those will come in subsequent CLs.